backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
sqlmodel==0.0.24
pymongo==4.11.3
httpx==0.28.1
python-dotenv==1.1.0
PyJWT==2.9.0
openpyxl==3.1.5
python-multipart==0.0.20
PyMuPDF==1.24.10
